taski/requirements_all.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.138
boto3-stubs==1.34.138
botocore==1.34.138
botocore-stubs==1.34.138
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
decorator==5.1.1
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-redis==5.4.0
django-storages==1.14.3
django-stubs==5.0.2
django-stubs-ext==5.0.2
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
drf-jwt==1.19.2
drf-spectacular==0.27.2
executing==2.0.1
factory-boy==3.3.0
Faker==26.0.0
flake8==7.1.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.10.1
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==8.2.2
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.3.0
txaio==23.1.1
types-awscrt==0.20.12
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240622
types-s3transfer==0.10.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==6.4.post2
